Add type's caller type parameter to function calls
Allow having exported functions in a type with a caller type parameter
that don't use the caller type parameter. This is done by specifying the
caller type parameter explicitly on every function call, so that the
compiler doesn't have to infer it, failing inference if there is no
caller parameter to infer from.